refactor: book 도메인 변경
