+++ src/main/java/com/sprint/sb06deokhugamteam01/dto/review/ReviewDto.java
@@ -1,6 +1,6 @@
 package com.sprint.sb06deokhugamteam01.dto.review;
 
-import com.sprint.sb06deokhugamteam01.domain.Book;
+import com.sprint.sb06deokhugamteam01.domain.book.Book;
 import com.sprint.sb06deokhugamteam01.domain.review.Review;
 import com.sprint.sb06deokhugamteam01.domain.User;
 import lombok.Builder;
